steamcompmgr: workaround bug with mangohud repaints when VRR is active
Ever since external overlay repaint logic was moved outside of
outdatedInteractiveFocus, drawing mangohud on top of the base plane
while VRR is enabled can cause the display to jump between different
refresh rates. This only seems to happen once a game runs at ~60% or
less of the display panel's maximum refresh rate, but it leads to very
choppy frame pacing and noticeable VRR flicker.

This choppy VRR effect can actually happen with every type of overlay,
including the Steam quick access menus, but mangohud is by far the biggest
culprit as it is the only visible overlay that is constantly being
drawn on top of the base plane.

This commit adds a workaround to avoid the worst flickering caused by
this phenomenon, but the underlying issue will still be present and should
be looked into.

Fixes: "steamcompmgr: Move outdatedInteractiveFocus to window"
Closes: #1513
Suggested-by: Antheas Kapenekakis <git@antheas.dev>