steamcompmgr: render dropdowns with negative coordinates, provided they are partially visible
The previous condition (strictly positive coordinates) was too strict,
it resulted in tall dropdowns not being rendered because they would have a
slightly negative Y coordinate.

An instance of this problem could be observed with the account creation screen
of Final Fantasy XIV Online (when using the webview2 launcher),
where the 'month' dropdown would be displayed correctly (as all 13 entries
fitted vertically), but the 'day' and 'year' dropdowns wouldn't be rendered
because they were taller, and their position ended up with a slightly negative
Y coordinate.